soccer_forecasting/requirements.txt
requests
joblib
sqlalchemy
scipy
urllib3
seaborn
statistics
Pillow
tabulate
tqdm
pandas
numpy

Pillow
bs4
highlight_text
matplotlib
mplsoccer

requests
retrying

PyYAML
sqlalchemy
unidecode

tabulate
urllib3
shapely
scikit-learn
psycopg2-binary

google-cloud-storage

pytz